# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface GenerateInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type e.g. image/jpeg */
  imageType: string;
  /** Target platform: Instagram, LinkedIn, YouTube */
  platform: string;
  /** Tone: Desi/Hinglish, Funny, Professional, Savage */
  tone: string;
}

export interface GeneratedCaption {
  text: string;
  hashtags: string[];
  cta: string;
}

export interface VisualAnalysis {
  sceneDescription: string;
  mood: string;
  keyObjects: string[];
  colorPalette: string[];
  humanCount: number;
}

export interface GenerateResult {
  captions: GeneratedCaption[];
  visualAnalysis: VisualAnalysis;
}

export interface RefineInput {
  captionText: string;
  platform: string;
  tone: string;
  hashtags: string[];
  cta: string;
}

export interface RefinedCaption {
  text: string;
  hashtags: string[];
  cta: string;
}

export interface SavedCaptionInput {
  text: string;
  hashtags: string[];
  cta: string;
  platform: string;
  tone: string;
  /**
     * Optional small base64 thumbnail
     * @nullable
     */
  imagePreviewBase64?: string | null;
}

export interface SavedCaption {
  id: number;
  text: string;
  hashtags: string[];
  cta: string;
  platform: string;
  tone: string;
  /** @nullable */
  imagePreviewBase64?: string | null;
  createdAt: string;
}

export interface StatCount {
  label: string;
  count: number;
}

export interface CaptionStats {
  totalSaved: number;
  byPlatform: StatCount[];
  byTone: StatCount[];
}

export type ListSavedCaptionsParams = {
platform?: string;
tone?: string;
limit?: number;
};