# lib/api-zod/src/generated/types/captionStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatCount } from './statCount';

export interface CaptionStats {
  totalSaved: number;
  byPlatform: StatCount[];
  byTone: StatCount[];
}