# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type e.g. image/jpeg */
  imageType: string;
  /** Target platform: Instagram, LinkedIn, YouTube */
  platform: string;
  /** Tone: Desi/Hinglish, Funny, Professional, Savage */
  tone: string;
}