# lib/api-zod/src/generated/types/generateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedCaption } from './generatedCaption';
import type { VisualAnalysis } from './visualAnalysis';

export interface GenerateResult {
  captions: GeneratedCaption[];
  visualAnalysis: VisualAnalysis;
}