# lib/api-zod/src/generated/types/generatedCaption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedCaption {
  text: string;
  hashtags: string[];
  cta: string;
}