# lib/api-zod/src/generated/types/refineInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RefineInput {
  captionText: string;
  platform: string;
  tone: string;
  hashtags: string[];
  cta: string;
}