# lib/api-zod/src/generated/types/savedCaption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavedCaption {
  id: number;
  text: string;
  hashtags: string[];
  cta: string;
  platform: string;
  tone: string;
  /** @nullable */
  imagePreviewBase64?: string | null;
  createdAt: Date;
}