# lib/api-zod/src/generated/types/savedCaptionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavedCaptionInput {
  text: string;
  hashtags: string[];
  cta: string;
  platform: string;
  tone: string;
  /**
     * Optional small base64 thumbnail
     * @nullable
     */
  imagePreviewBase64?: string | null;
}