# lib/api-zod/src/generated/types/visualAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CaptionCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VisualAnalysis {
  sceneDescription: string;
  mood: string;
  keyObjects: string[];
  colorPalette: string[];
  humanCount: number;
}